tools/requirements.txt
cliff==3.5.0
PyYAML==5.3.1
Jinja2==2.11.2
jsonschema==3.2.0